+++ src/components/survey/SurveySingleItemView/stories/SingleChoice.stories.tsx
@@ -82,3 +82,4 @@ export const ExampleWithMissingOptions = () => <SurveySingleItemView
   invalidWarning={invalidWarning}
   showKeys={true}
 />
+